refactor: 출발,도착역을 반대로 입력해도 수행되도록 수정
